API 1.6.0 - Add server response signature option
This API update is made to work with clients using version 5.7.0 and above. It is STILL COMPATIBLE with older clients! :)

This update is made to replace the "SSL Pinning", by signing every response of the server (when the clients are in version 5.7.0+) with SHA256 and a private RSA key. Every client will be able to check the signature, using the server's public key.